+++ projects/store/src/decorators/index.ts
@@ -1,4 +1,4 @@
 export * from './dispatch';
-export * from './ngrx-select';
+export { NgrxSelectModule, NgrxSelect as ɵNgrxSelect } from './ngrx-select';
 export * from './pluck';
 export * from './select';