+++ projects/store/src/store.ts
@@ -1,18 +1,4 @@
-export {
-  Dispatch,
-  Select,
-  NgrxSelect as ɵNgrxSelect,
-  NgrxSelectModule,
-  Pluck
-} from './decorators/index';
-
-export {
-  NgLetDirective,
-  NgLetModule,
-  RouterLinkActiveMatchModule,
-  RouterLinkActiveMatch
-} from './directives/index';
-
-export { untilDestroy, destroy$ as ɵdestroy$ } from './operators/index';
-
-export { PushPipe, PushPipeModule } from './pipes/index';
+export * from './decorators/index';
+export * from './directives/index';
+export * from './operators/index';
+export * from './pipes/index';